Nearly complete re-write of the game. *BROKEN*
Removed Cell entirely. Cells are used, but as a part of Block.

Rotation is working.
Movement and screen boundaries are working - a block can now never leave
the screen, even partially.

However, when we went to re-implement block collision, it gave us some
major errors. Currently do not know the source or means of fixing.